EXAMPLE: Motivating example for OTA-1531
Illustrates the caveat of the current approach to determining which gates are enabled and disabled in the cluster. The current approach performs this initialization late in the execution (once the initial payyload is fully loaded) so there is an awkward period on CVO startup where enabled gates cannot be checked and the programmer cannot gate behavior.