refactor(payload): separate arch load from metadata and runtime
The architecture is either read from release metadata when present (and in that case it must be a string and must be multi) or determined from runtime in all other cases. Refactor `loadReleaseFromMetadata` so that it only really reads the metadata, and move the "when metadata does not have architecture, read from runtime" logic to `loadPayloadMetadata` above (the `loadReleaseFromMetadata` caller). This will allow to reuse the release metadata loading code without involving more logic than necessary.

This also simplifies `loadReleaseFromMetadata` signature, improving readability.

Also turn some nested conditionals into guard clauses for readability.