cvo: determine featureSet and CVO gates early
Extracting config.openshift.io informer factories allows to obtain the
`FeatureGate` lister early. We need to start the the config informer
factory for that but that is fine, later code can create new informers
as long as it calls `Start()` again. Retrieved `FeatureGate` can then be
used to read the featureSet enabled in the cluster and also the enabled
and disabled CVO feature gates.

This is basically identical to what CVO currently does later in
`Context.InitializeFromPayload`. For now _that_ code drives the actual
CVO behavior, not the values discovered by the early code. The change to
switch the CVO behavior from respecting the late code to the early code
will follow; this commit just sets the stage for that and allows testing
the early informer start.