Clean up the condition with type=ImplicitlyEnabled
This commit addresses the issue found in the last round of pre-merge
test [1].

This commit could be reverted after maximal 2 releases.
(It is alright to keep it too)
Assume this PR gets merged to main which is ATM dev branch: 4.20.
When the main branches points to 4.22, all the clusters affetecd
by the bug (originally 4.19 and 4.20) should be fixed when they
are upgraded to 4.20/4.21. Thus any 4.22 cluster should not be
affected.

[1]. https://issues.redhat.com/browse/OCPBUGS-56114?focusedId=27251461&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-27251461