pkg/cincinnati: Centralize release metadata parsing
Instead of carring similar code in the payload and Cincinnati
packages, centralize in one place to make maintenance easier.  And
with the logic centralized, I've also put some time into hardening the
channel parsing, to grumble about some possible issues.

For payload loading, errors get a logged warning, but are not fatal.
Having the CVO come up with logged warnings still allows cluster
admins to update into a fix.  But a crash-looping CVO would not notice
ClusterVersion spec.desiredUpdate changes or be able to roll out a
requested update into a fix.  So this branch processes as much of the
metadata as it can (to not let, for example, an invalid URL block us
from loading architecture information), logs the warning about what it
failed to parse, and continues on.

For Cincinnati processing, errors are fatal.  Cluster admins can take
the ClusterVersion RetrievedUpdates=False message and complain to
their Update Service maintainers, who can fix the metadata.