config: Conditionally initialize configuration controller
The information regarding enabled feature gates is now available
sooner in the code. Utilize this information to conditionally initialize
the configuration controller at a one specific place. If no configuration
controller is needed, this information is clearly stated by
the noninitialized configuration field.

The subsequent code can then easily check for existence of the configuration
using a nil.