config: Introduce configuration provider interface
Introduce a configuration provider interface. The goal of the interface
is to abstract the underlying system of fetching the CVO configuration.
Be it a Kubernetes API server, or a configuration file. For the moment,
the Kubernetes API server is the sole provider.

Refactor the code to adhere to the new logic.

Apply a default configuration on a NotFound error (was previously noted
as a TODO in the source code).