test: Rename test to comply with OTE integration guide
Per the integration guide for OTE, each test must have a clearly defined
owner (by using a [Jira:Component] tag) or an entry in the
ci-test-mapping repository. Naming the owner as
"Cluster Version Operator" with spaces, as this is the format used to
distinguish our component in the mapping repository [1].

Renaming tests should not be done without a proper setup, as this can
lose a test's history. However, this test was never executed, and it
acts as a signal of the OTE integration.

[1]: https://github.com/openshift-eng/ci-test-mapping/blob/b2ca5844ead12ede1cb72cae34321f3215beb6e3/pkg/components/clusterversionoperator/component.go#L14