availableupdates: Refactor getting desired/current architecture
The commit does change the default behaviour of the
`getDesiredArchitecture` method. However, the method is only used once
in the `syncAvailableUpdates` method and nowhere else. The commit adds
the subsequent logic for evaluating a desired architecture to the method
itself and implements the `getCurrentArchitecture` method.

The goal is to introduces "getters" for such values where the
unified capitalization is enforced, and their return values are of the
same nature (e.g., "Multi", "amd64", ...).