test: Add quotation marks to sanity test's component name
The CVO jira component contains spaces in its name. Add apostrophes
to make the component name more robust.

Both options (with and without quotation marks) can be seen in OpenShift
repositories. This is done due to functionality concerns to be rather on
the safer side.

Renaming tests should not be done without a proper setup, as this can
lose a test's history. However, this test was never executed, and it
acts as a signal of the OTE integration.