test: Manually rename existing test metadata
Such changes should not be done manually after the extension is
integrated!

Renaming tests should not be done without a proper setup, as this can
lose a test's history. However, this test was never executed, and it
acts as a signal of the OTE integration.

Doing this change manually to not unnecessarily clutter the code with
setting the original name of a test, which was never executed.