add dedicated service account to cluster-version-operator- , version- , and update-payload- pods
This change adds a dedicated service account to the pods mentioned above. The ClusterRoleBinding `cluster-version-operator-1`
bind the necessary `cluster-admin` ClusterRole to the `cluster-version-operator` ServiceAccount.

The ClusterRoleBinding `cluster-version-operator` is still required for e2e-agnostic-ovn-upgrade-out-of-change test
to pass, as the `default` ServiceAccount is still in use here. I suspect we can remove this role binding once this
PR merges.

update-payload pod just needs its own service account, no additional permissions required.

TestRenderManifestDeployment_expected_cvo_deployment.yaml has a bespoke service account added to match the CVO
Deployment, so that the tests pass for this change.